Add files via upload
fullscript.py is the most recent version of the script. It finds the clinical variants that occur in strongly, weakly, and fully conserved regions. It will continue to be optimized in order to find variants in non-conserved regions and to add a line to the MSA that annotates where the mutations occur. parse_clin_var.py is another script that parses the clinical variant file from NCBI in order to allow fullscript.py to run faster.